in workspace, added docs for api available in lua batch scripts in spel-katalog-build, spel-katalog-settings, added methods to access settings as a hash map from str to strings. in spel-katalog-batch, added settings access for lua scripts
