+++ lua_api.md
@@ -1,0 +1,14 @@
+# Lua API for batch scripts
+
+## `dbg(Any..) -> Any..`
+Debug print and return all passed values.
+
+## `loadYaml(path: String)`
+Load yaml at path into a lua value.
+
+## `None`
+A null value separate from lua nil, used for optional
+values.
+
+## `settings`
+Current settings as a table.
