+++ spel-katalog-batch/Cargo.toml
@@ -10,6 +10,7 @@ iced_highlighter.workspace = true
 log.workspace = true
 mlua = { workspace = true, features = ["serde", "lua54", "vendored"]}
 rfd = { workspace = true, features = ["xdg-portal","tokio"] }
+rusqlite.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 spel-katalog-common.workspace = true
